[BugFix]Fix some bug about binlog config (#16534)
- fix bug, can not load editlog of `modify_binlog` and `modify binlogAvailableVersion`
- fix bug, if only `binlog_enable` is set , then binlog config can not set successfully
- change the logic of `setBinlogConfig`
- fix the logic of `checkIsPartitionChanged`
- optimize some code, such as the format of log info

+++ fe/fe-core/src/main/java/com/starrocks/common/Config.java
@@ -1992,5 +1992,5 @@ public class Config extends ConfigBase {
     public static long binlog_ttl_second = 60 * 30; // 30min
 
     @ConfField
-    public static long binlog_max_size = Integer.MAX_VALUE; // no limit
+    public static long binlog_max_size = Long.MAX_VALUE; // no limit
 }
